Type Definition atsamd11d::adc::inputctrl::MUXPOS_R

type MUXPOS_R = R<u8, MUXPOS_A>;

Reader of field MUXPOS

Implementations

impl MUXPOS_R

pub fn variant(&self) -> Variant<u8, MUXPOS_A>

Get enumerated values variant

pub fn is_pin0(&self) -> bool

Checks if the value of the field is PIN0

pub fn is_pin1(&self) -> bool

Checks if the value of the field is PIN1

pub fn is_pin2(&self) -> bool

Checks if the value of the field is PIN2

pub fn is_pin3(&self) -> bool

Checks if the value of the field is PIN3

pub fn is_pin4(&self) -> bool

Checks if the value of the field is PIN4

pub fn is_pin5(&self) -> bool

Checks if the value of the field is PIN5

pub fn is_pin6(&self) -> bool

Checks if the value of the field is PIN6

pub fn is_pin7(&self) -> bool

Checks if the value of the field is PIN7

pub fn is_pin8(&self) -> bool

Checks if the value of the field is PIN8

pub fn is_pin9(&self) -> bool

Checks if the value of the field is PIN9

pub fn is_pin10(&self) -> bool

Checks if the value of the field is PIN10

pub fn is_pin11(&self) -> bool

Checks if the value of the field is PIN11

pub fn is_pin12(&self) -> bool

Checks if the value of the field is PIN12

pub fn is_pin13(&self) -> bool

Checks if the value of the field is PIN13

pub fn is_pin14(&self) -> bool

Checks if the value of the field is PIN14

pub fn is_pin15(&self) -> bool

Checks if the value of the field is PIN15

pub fn is_pin16(&self) -> bool

Checks if the value of the field is PIN16

pub fn is_pin17(&self) -> bool

Checks if the value of the field is PIN17

pub fn is_pin18(&self) -> bool

Checks if the value of the field is PIN18

pub fn is_pin19(&self) -> bool

Checks if the value of the field is PIN19

pub fn is_temp(&self) -> bool

Checks if the value of the field is TEMP

pub fn is_bandgap(&self) -> bool

Checks if the value of the field is BANDGAP

pub fn is_scaledcorevcc(&self) -> bool

Checks if the value of the field is SCALEDCOREVCC

pub fn is_scalediovcc(&self) -> bool

Checks if the value of the field is SCALEDIOVCC

pub fn is_dac(&self) -> bool

Checks if the value of the field is DAC