Type Definition atsamd11d::nvmctrl::ctrlb::SLEEPPRM_R

type SLEEPPRM_R = R<u8, SLEEPPRM_A>;

Reader of field SLEEPPRM

Implementations

impl SLEEPPRM_R

pub fn variant(&self) -> Variant<u8, SLEEPPRM_A>

Get enumerated values variant

pub fn is_wakeonaccess(&self) -> bool

Checks if the value of the field is WAKEONACCESS

pub fn is_wakeupinstant(&self) -> bool

Checks if the value of the field is WAKEUPINSTANT

pub fn is_disabled(&self) -> bool

Checks if the value of the field is DISABLED