Type Definition atsamd11d::port::pmux0_::PMUXO_R

type PMUXO_R = R<u8, PMUXO_A>;

Reader of field PMUXO

Implementations

impl PMUXO_R

pub fn variant(&self) -> Variant<u8, PMUXO_A>

Get enumerated values variant

pub fn is_a(&self) -> bool

Checks if the value of the field is A

pub fn is_b(&self) -> bool

Checks if the value of the field is B

pub fn is_c(&self) -> bool

Checks if the value of the field is C

pub fn is_d(&self) -> bool

Checks if the value of the field is D

pub fn is_e(&self) -> bool

Checks if the value of the field is E

pub fn is_f(&self) -> bool

Checks if the value of the field is F

pub fn is_g(&self) -> bool

Checks if the value of the field is G

pub fn is_h(&self) -> bool

Checks if the value of the field is H