atsamd11d::tcc0::ctrlbclr
Type Alias IdxcmdW

pub type IdxcmdW<'a, REG> = FieldWriter<'a, REG, 2, Idxcmdselect, Safe>;

Field IDXCMD writer - Ramp Index Command

Aliased Type

struct IdxcmdW<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> IdxcmdW<'a, REG>
where REG: Writable + RegisterSpec, REG::Ux: From<u8>,

pub fn disable(self) -> &'a mut W<REG>

Command disabled: Index toggles between cycles A and B

pub fn set_(self) -> &'a mut W<REG>

Set index: cycle B will be forced in the next cycle

pub fn clear(self) -> &'a mut W<REG>

Clear index: cycle A will be forced in the next cycle

pub fn hold(self) -> &'a mut W<REG>

Hold index: the next cycle will be the same as the current cycle