Type Definition atsamd11d::tcc0::evctrl::EVACT0_R

type EVACT0_R = R<u8, EVACT0_A>;

Reader of field EVACT0

Implementations

impl EVACT0_R

pub fn variant(&self) -> Variant<u8, EVACT0_A>

Get enumerated values variant

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF

pub fn is_retrigger(&self) -> bool

Checks if the value of the field is RETRIGGER

pub fn is_countev(&self) -> bool

Checks if the value of the field is COUNTEV

pub fn is_start(&self) -> bool

Checks if the value of the field is START

pub fn is_inc(&self) -> bool

Checks if the value of the field is INC

pub fn is_count(&self) -> bool

Checks if the value of the field is COUNT

pub fn is_fault(&self) -> bool

Checks if the value of the field is FAULT