Type Definition atsamd11d::tcc0::evctrl::EVACT1_R

type EVACT1_R = R<u8, EVACT1_A>;

Reader of field EVACT1

Implementations

impl EVACT1_R

pub fn variant(&self) -> EVACT1_A

Get enumerated values variant

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF

pub fn is_retrigger(&self) -> bool

Checks if the value of the field is RETRIGGER

pub fn is_dir(&self) -> bool

Checks if the value of the field is DIR

pub fn is_stop(&self) -> bool

Checks if the value of the field is STOP

pub fn is_dec(&self) -> bool

Checks if the value of the field is DEC

pub fn is_ppw(&self) -> bool

Checks if the value of the field is PPW

pub fn is_pwp(&self) -> bool

Checks if the value of the field is PWP

pub fn is_fault(&self) -> bool

Checks if the value of the field is FAULT