Type Definition atsamd11d::tcc0::fctrla::CAPTURE_R

type CAPTURE_R = R<u8, CAPTURE_A>;

Reader of field CAPTURE

Implementations

impl CAPTURE_R

pub fn variant(&self) -> CAPTURE_A

Get enumerated values variant

pub fn is_disable(&self) -> bool

Checks if the value of the field is DISABLE

pub fn is_capt(&self) -> bool

Checks if the value of the field is CAPT

pub fn is_captmin(&self) -> bool

Checks if the value of the field is CAPTMIN

pub fn is_captmax(&self) -> bool

Checks if the value of the field is CAPTMAX

pub fn is_locmin(&self) -> bool

Checks if the value of the field is LOCMIN

pub fn is_locmax(&self) -> bool

Checks if the value of the field is LOCMAX

pub fn is_deriv0(&self) -> bool

Checks if the value of the field is DERIV0

pub fn is_captmark(&self) -> bool

Checks if the value of the field is CAPTMARK