atsamd11d::usb::device

Module intenset

DEVICE Device Interrupt Enable Set

Structs

IntensetSpec

DEVICE Device Interrupt Enable Set

Type Aliases

EorsmR

Field EORSM reader - End Of Resume Interrupt Enable

EorsmW

Field EORSM writer - End Of Resume Interrupt Enable

EorstR

Field EORST reader - End of Reset Interrupt Enable

EorstW

Field EORST writer - End of Reset Interrupt Enable

LpmnyetR

Field LPMNYET reader - Link Power Management Not Yet Interrupt Enable

LpmnyetW

Field LPMNYET writer - Link Power Management Not Yet Interrupt Enable

LpmsuspR

Field LPMSUSP reader - Link Power Management Suspend Interrupt Enable

LpmsuspW

Field LPMSUSP writer - Link Power Management Suspend Interrupt Enable

MsofR

Field MSOF reader - Micro Start of Frame Interrupt Enable in High Speed Mode

MsofW

Field MSOF writer - Micro Start of Frame Interrupt Enable in High Speed Mode

R

Register INTENSET reader

RamacerR

Field RAMACER reader - Ram Access Interrupt Enable

RamacerW

Field RAMACER writer - Ram Access Interrupt Enable

SofR

Field SOF reader - Start Of Frame Interrupt Enable

SofW

Field SOF writer - Start Of Frame Interrupt Enable

SuspendR

Field SUSPEND reader - Suspend Interrupt Enable

SuspendW

Field SUSPEND writer - Suspend Interrupt Enable

UprsmR

Field UPRSM reader - Upstream Resume Interrupt Enable

UprsmW

Field UPRSM writer - Upstream Resume Interrupt Enable

W

Register INTENSET writer

WakeupR

Field WAKEUP reader - Wake Up Interrupt Enable

WakeupW

Field WAKEUP writer - Wake Up Interrupt Enable