Crate atsamd11d

Peripheral access API for ATSAMD11D microcontrollers (generated using svd2rust v0.33.5 ( ))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Re-exports

pub use self::pac0 as pac1;

pub use self::pac0 as pac2;

pub use self::port as port_iobus;

pub use self::sercom0 as sercom1;

pub use self::sercom0 as sercom2;

pub use self::tc1 as tc2;

Modules

ac

Analog Comparators

adc

Analog Digital Converter

dac

Digital Analog Converter

dmac

Direct Memory Access Controller

dsu

Device Service Unit

eic

External Interrupt Controller

evsys

Event System Interface

gclk

Generic Clock Generator

generic

Common register and bit access and modify traits

hmatrix

HSB Matrix

mtb

Cortex-M0+ Micro-Trace Buffer

nvmctrl

Non-Volatile Memory Controller

pac0

Peripheral Access Controller 0

pm

Power Manager

port

Port Module

rtc

Real-Time Counter

sercom0

Serial Communication Interface 0

sysctrl

System Control

tc1

Basic Timer Counter 1

tcc0

Timer Counter Control

usb

Universal Serial Bus

wdt

Watchdog Timer

Structs

Ac

Analog Comparators

Adc

Analog Digital Converter

CBP

Cache and branch predictor maintenance operations

CPUID

CPUID

CorePeripherals

Core peripherals

DCB

Debug Control Block

DWT

Data Watchpoint and Trace unit

Dac

Digital Analog Converter

Dmac

Direct Memory Access Controller

Dsu

Device Service Unit

Eic

External Interrupt Controller

Evsys

Event System Interface

FPB

Flash Patch and Breakpoint unit

Gclk

Generic Clock Generator

Hmatrix

HSB Matrix

ITM

Instrumentation Trace Macrocell

MPU

Memory Protection Unit

Mtb

Cortex-M0+ Micro-Trace Buffer

NVIC

Nested Vector Interrupt Controller

Nvmctrl

Non-Volatile Memory Controller

Pac0

Peripheral Access Controller 0

Pac1

Peripheral Access Controller 1

Pac2

Peripheral Access Controller 2

Peripherals

All the peripherals.

Pm

Power Manager

Port

Port Module

PortIobus

Port Module (IOBUS)

Rtc

Real-Time Counter

SCB

System Control Block

SYST

SysTick: System Timer

Sercom0

Serial Communication Interface 0

Sercom1

Serial Communication Interface 1

Sercom2

Serial Communication Interface 2

Sysctrl

System Control

TPIU

Trace Port Interface Unit

Tc1

Basic Timer Counter 1

Tc2

Basic Timer Counter 2

Tcc0

Timer Counter Control

Usb

Universal Serial Bus

Wdt

Watchdog Timer

Enums

Interrupt

Enumeration of all the interrupts.

Constants

NVIC_PRIO_BITS

Number available in the NVIC for configuring priority